+++ lmcache/sdk/__init__.py
@@ -2,9 +2,6 @@
 """Public LMCache SDK helpers."""
 
 # First Party
-from lmcache.sdk.kvcache import KVCacheSDKError, LMCacheKVCacheContext
+from lmcache.sdk import batch, context, kvcache, qcache, stream
 
-__all__ = [
-    "KVCacheSDKError",
-    "LMCacheKVCacheContext",
-]
+__all__ = ["batch", "context", "kvcache", "qcache", "stream"]
