+++ lmcache/sdk/stream.py
@@ -7,7 +7,7 @@ Public API for LMCacheStream, a wrapper of a logical request going through the S
 from __future__ import annotations
 
 # Standard
-from collections.abc import Callable, Iterable, Sequence
+from collections.abc import Iterable, Sequence
 from dataclasses import dataclass, field
 from typing import Any, Protocol
 import time
@@ -18,7 +18,11 @@ import torch
 
 # First Party
 from lmcache.logging import init_logger
-import lmcache.sdk.kvcache as lmc_sdk
+from lmcache.sdk.context import (
+    LMCacheSDKCacheKind,
+    LMCacheSDKContext,
+    ModifyFnType,
+)
 
 logger = init_logger(__name__)
 
@@ -86,7 +90,7 @@ class PostCompletion(Protocol):
 
 
 def create_request(
-    ctx: lmc_sdk.LMCacheKVCacheContext,
+    contexts: Iterable[LMCacheSDKContext],
     post_completion: PostCompletion,
     prompt_token_ids: Sequence[int],
     cache_salt: str = "",
@@ -94,7 +98,7 @@ def create_request(
     """Create an LMCacheStream for a new request.
 
     Args:
-        ctx: The LMCache SDK context used for retrieve/store.
+        contexts: The LMCache SDK contexts used for retrieve/store.
         post_completion: Callable that submits a request to the engine.
         prompt_token_ids: Initial prompt token ids.
         cache_salt: Per-user isolation salt, or empty string.
@@ -103,43 +107,18 @@ def create_request(
         A new LMCacheStream.
     """
     return LMCacheStream(
-        ctx=ctx,
+        contexts=contexts,
         post_completion=post_completion,
         prompt_token_ids=prompt_token_ids,
         cache_salt=cache_salt,
     )
 
 
-def generate(
-    stream: LMCacheStream,
-    sampling_params: dict[str, Any],
-    suffix_tokens: Sequence[int] = (),
-) -> StreamPerfMetrics:
-    """Run/continue a stream's generate. See LMCacheStream.generate."""
-    return stream.generate(sampling_params, suffix_tokens)
-
-
-def get(
-    stream: LMCacheStream, timeout: float = 30.0, poll_interval: float = 0.2
-) -> torch.Tensor:
-    """Retrieve the cached KV for a stream. See LMCacheStream.retrieve_kv."""
-    return stream.retrieve_kv(timeout=timeout, poll_interval=poll_interval)
-
-
-def update(
-    stream: LMCacheStream,
-    kv: torch.Tensor,
-    tokens: Sequence[int],
-) -> None:
-    """Store edited KV into a stream. See LMCacheStream.update_kv."""
-    stream.update_kv(kv, tokens)
-
-
 class LMCacheStream:
     """Handle for one logical request spanning multiple inference passes.
 
     Args:
-        ctx: The LMCache SDK context used for retrieve/store.
+        contexts: The LMCache SDK contexts used for retrieve/store.
         post_completion: Callable for submitting request to inference engine.
         prompt_token_ids: Initial prompt token ids.
         cache_salt: Per-user isolation salt, or empty string.
@@ -147,7 +126,7 @@ class LMCacheStream:
 
     def __init__(
         self,
-        ctx: lmc_sdk.LMCacheKVCacheContext,
+        contexts: Iterable[LMCacheSDKContext],
         post_completion: PostCompletion,
         prompt_token_ids: Sequence[int],
         cache_salt: str = "",
@@ -158,7 +137,9 @@ class LMCacheStream:
         the KV, starting as the prompt), done (the EOS flag), and internal
         output history / suffix-token bookkeeping.
         """
-        self.ctx = ctx
+        self._contexts: dict[LMCacheSDKCacheKind, LMCacheSDKContext] = {}
+        for ctx in contexts:
+            self._contexts[ctx.kind] = ctx
         self.post_completion = post_completion
         self.cache_salt = cache_salt
         self.tokens: list[int] = list(prompt_token_ids)
@@ -168,26 +149,32 @@ class LMCacheStream:
         self._stream_id: str = str(uuid.uuid4())
         self._suffix_tokens: list[int] = []
 
+    @property
     def stream_id(self) -> str:
         """Return the unique stream id."""
         return self._stream_id
 
+    @property
     def suffix_tokens(self) -> list[int]:
         """Return the suffix tokens to be appended to the prompt."""
         return self._suffix_tokens
 
+    @property
     def decoded_tokens(self) -> int:
         """Return cumulative tokens decoded across all segments so far."""
         return self._decoded
 
+    @property
     def output_text(self) -> str:
         """Return the concatenated generated text across all segments."""
         return "".join(self._text_parts)
 
+    @property
     def output_tokens(self) -> list[int]:
         """Return the concatenated generated tokens across all segments."""
         return self.tokens
 
+    @property
     def is_done(self) -> bool:
         """Return True if the stream has finished generating."""
         return self.done
@@ -253,34 +240,44 @@ class LMCacheStream:
             ttft=time_between_tokens[0] if time_between_tokens else 0.0,
         )
 
-    def retrieve_kv(
-        self, timeout: float = 30.0, poll_interval: float = 0.2
+    def retrieve(
+        self,
+        kind: LMCacheSDKCacheKind,
+        timeout: float = 30.0,
+        poll_interval: float = 0.2,
     ) -> torch.Tensor:
-        """Retrieve the cached KV for the current tokens, polling until ready.
+        """Retrieve the cached tensor for the current tokens, polling until
+        ready.
 
         Args:
-            timeout: Max seconds to wait for the cached KV to appear.
+            kind: The type of cache to retrieve.
+            timeout: Max seconds to wait for the cached tensor to appear.
             poll_interval: Seconds between retrieve attempts.
 
         Returns:
-            The cached KV as a [2, L, hit_tokens, D] tensor (chunk-aligned).
+            The cached tensor (chunk-aligned).
+            KV shape is [2, L, T, D]. Q shape is [1, L, T, D].
 
         Raises:
-            LMCacheStreamError: If no cached KV is available within timeout.
+            LMCacheStreamError: If no cached tensor is available within timeout.
         """
+        ctx = self._contexts.get(kind)
+        if not ctx:
+            raise LMCacheStreamError(f"no context available for cache kind {kind}")
         deadline = time.perf_counter() + timeout
-        kv = lmc_sdk.retrieve(self.ctx, self.tokens, self.cache_salt)
-        while kv is None and time.perf_counter() < deadline:
+        tensor = ctx.retrieve(self.tokens, self.cache_salt)
+        while tensor is None and time.perf_counter() < deadline:
             time.sleep(poll_interval)
-            kv = lmc_sdk.retrieve(self.ctx, self.tokens, self.cache_salt)
-        if kv is None:
+            tensor = ctx.retrieve(self.tokens, self.cache_salt)
+        if tensor is None:
             raise LMCacheStreamError(
-                f"no cached KV for {self.stream_id} after {timeout:.0f}s"
+                f"no cached {kind} for {self.stream_id} after {timeout:.0f}s"
             )
-        return kv
+        return tensor
 
-    def update_kv(
+    def update(
         self,
+        kind: LMCacheSDKCacheKind,
         kv: torch.Tensor,
         tokens: Sequence[int],
     ) -> None:
@@ -293,7 +290,10 @@ class LMCacheStream:
             kv: The edited KV tensor to store, shape [2, L, T, D].
             tokens: Token ids the KV corresponds to (T must match kv.shape[2]).
         """
-        if not lmc_sdk.store(self.ctx, kv, tokens, self.cache_salt):
+        ctx = self._contexts.get(kind)
+        if not ctx:
+            raise LMCacheStreamError(f"no context available for cache kind {kind}")
+        if not ctx.store(kv, tokens, self.cache_salt):
             logger.warning(
                 "store reported edited KV already cached for stream %s",
                 self.stream_id,
@@ -303,7 +303,7 @@ class LMCacheStream:
 
     def modify_kv(
         self,
-        fn: Callable[[torch.Tensor, Sequence[int]], tuple[torch.Tensor, Sequence[int]]],
+        fn: ModifyFnType,
         timeout: float = 30.0,
         poll_interval: float = 0.2,
     ) -> None:
@@ -314,13 +314,21 @@ class LMCacheStream:
         applies fn to the cached prefix, and stores the result via update_kv.
 
         Args:
-            fn: KV editor given (kv[2,L,T,D], tokens[:cached_len]) and returning
-                (new_kv, new_tokens) for the edited prefix.
+            fn: KV editor given Mapping[LMCacheSDKCacheKind, torch.Tensor]
+                for each cache kind used in the modification algorithm,
+                returning (new_kv, new_tokens) for the edited prefix.
         """
-        kv = self.retrieve_kv(timeout=timeout, poll_interval=poll_interval)
-        cached_len = kv.shape[2]
+        cached_len = len(self.tokens)
+        tensors: dict[LMCacheSDKCacheKind, torch.Tensor] = {}
+        for ctx in self._contexts.values():
+            tensors[ctx.kind] = self.retrieve(
+                kind=ctx.kind, timeout=timeout, poll_interval=poll_interval
+            )
+            if ctx.kind == LMCacheSDKCacheKind.KV:
+                cached_len = tensors[ctx.kind].shape[2]
+
         # Tokens past the cached KV: the remainder of chunks that retrieve()
         # (chunk-aligned) didn't return.
         self._suffix_tokens = list(self.tokens[cached_len:]) + self._suffix_tokens
-        new_kv, new_tokens = fn(kv, self.tokens[:cached_len])
-        self.update_kv(new_kv, new_tokens)
+        new_kv, new_tokens = fn(tensors, self.tokens[:cached_len])
+        self.update(kind=LMCacheSDKCacheKind.KV, kv=new_kv, tokens=new_tokens)
